gulp-styledocco should be able to take a directory
In before behavior, gulp-styledocco cannot take a directory.
Because `directory.isNull()` is true, and `firstFile` becomes null.
And when gulp-styledocco take multiple files, it care only
the base path of the first file. It is dangerous behavior,
because the file order higly depend to the glob-stream implementation.

For these reasons, gulp-styledocco should take a directory.
And when multiple files arrived, it should pass all files
to the styledocco command. The command can take multiple files.